requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.0
fsspec==2025.12.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sklearn==0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.4.1
torch-directml==0.2.5.dev240914
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.1
Werkzeug==3.1.4
